pyproject.toml
[tool.poetry]
name = "geonetwork_resources"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
pydantic = "^1.10.2"
PyYAML = "^6.0"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
